[BPK-4331] Migrate R-Z stories to new format
