+++ .storybook/config.js
@@ -122,7 +122,6 @@ configure(() => {
   require('../packages/bpk-component-text/stories');
   require('../packages/bpk-component-textarea/stories');
   require('../packages/bpk-component-ticket/stories');
-  require('../packages/bpk-component-tile/stories');
   require('../packages/bpk-component-tooltip/stories');
   require('../packages/bpk-mixins/stories');
 }, module);
