+++ packages/bpk-component-radio/stories.js
@@ -18,103 +18,21 @@
 
 /* @flow strict */
 
-import React from 'react';
 import { storiesOf } from '@storybook/react';
-import { action } from 'bpk-storybook-utils';
-import { colorSkyGrayTint01, spacingBase } from 'bpk-tokens/tokens/base.es6';
 
-import BpkRadio from './index';
-
-const loremIpsum = `Lorem ipsum dolor sit amet, consectetur adipisicing elit. Dolorem dolores doloremque, expedita
-quaerat temporibus ipsam, ut, ipsa, velit sed assumenda suscipit dolore quod similique delectus numquam neque!
-Nesciunt, voluptate, illo.`;
+import {
+  DefaultExample,
+  Multiline,
+  Invalid,
+  White,
+  DisabledChecked,
+  DisabledUnchecked,
+} from './examples';
 
 storiesOf('bpk-component-radio', module)
-  .add('Checked', () => (
-    <BpkRadio
-      id="checked"
-      name="checked"
-      label="Return"
-      onChange={action('radio changed')}
-      checked
-    />
-  ))
-  .add('Unchecked', () => (
-    <BpkRadio
-      id="unchecked"
-      name="unchecked"
-      label="Return"
-      onChange={action('radio changed')}
-    />
-  ))
-  .add('Multi line', () => (
-    <BpkRadio
-      id="multi_line"
-      name="multi_line"
-      label={loremIpsum}
-      onChange={action('radio changed')}
-    />
-  ))
-  .add('Invalid', () => (
-    <BpkRadio
-      id="invalid"
-      name="invalid"
-      label="Return"
-      onChange={action('radio changed')}
-      valid={false}
-    />
-  ))
-  .add('White (Checked)', () => (
-    <div style={{ backgroundColor: colorSkyGrayTint01, padding: spacingBase }}>
-      <BpkRadio
-        id="checked"
-        name="checked"
-        label="Return"
-        onChange={action('radio changed')}
-        white
-        checked
-      />
-    </div>
-  ))
-  .add('White (Unchecked)', () => (
-    <div style={{ backgroundColor: colorSkyGrayTint01, padding: spacingBase }}>
-      <BpkRadio
-        id="unchecked"
-        name="unchecked"
-        label="Return"
-        onChange={action('radio changed')}
-        white
-      />
-    </div>
-  ))
-  .add('White (Invalid)', () => (
-    <div style={{ backgroundColor: colorSkyGrayTint01, padding: spacingBase }}>
-      <BpkRadio
-        id="invalid"
-        name="invalid"
-        label="Return"
-        onChange={action('radio changed')}
-        valid={false}
-        white
-      />
-    </div>
-  ))
-  .add('Disabled (Checked)', () => (
-    <BpkRadio
-      id="disabled_checked"
-      name="disabled_checked"
-      label="Return"
-      onChange={action('radio changed')}
-      checked
-      disabled
-    />
-  ))
-  .add('Disabled (Unchecked)', () => (
-    <BpkRadio
-      id="disabled"
-      name="disabled"
-      label="Return"
-      onChange={action('radio changed')}
-      disabled
-    />
-  ));
+  .add('Default', DefaultExample)
+  .add('Multi line', Multiline)
+  .add('Invalid', Invalid)
+  .add('White', White)
+  .add('Disabled (Checked)', DisabledChecked)
+  .add('Disabled (Unchecked)', DisabledUnchecked);
