+++ packages/bpk-component-rating/examples.scss
