+++ packages/bpk-component-rating/stories.js
@@ -17,294 +17,27 @@
  */
 /* @flow strict */
 
-import React from 'react';
 import { storiesOf } from '@storybook/react';
-import { cssModules } from 'bpk-react-utils';
 
-import STYLES from './stories.scss';
-
-import BpkRating, { RATING_SIZES, RATING_TYPES } from './index';
-
-const getClassName = cssModules(STYLES);
+import {
+  DefaultExample,
+  Vertical,
+  LargeSize,
+  LargeVertical,
+  SmallSize,
+  SmallVertical,
+  TitleOnly,
+  PillRatings,
+  PillTitleOnly,
+} from './examples';
 
 storiesOf('bpk-component-rating', module)
-  .add('Default', () => (
-    <div>
-      <BpkRating
-        ariaLabel="9 Excellent would recommend"
-        title="Excellent"
-        subtitle="This place was amazing"
-        value={9}
-      />
-      <br />
-      <BpkRating
-        ariaLabel="6.7 Average might recommend"
-        title="Average"
-        subtitle="Might recommend"
-        value={6.7}
-      />
-      <br />
-      <BpkRating
-        ariaLabel="2.3 Bad avoid here"
-        title="Bad"
-        subtitle="Avoid here"
-        value={2.3}
-      />
-    </div>
-  ))
-  .add('Vertical', () => (
-    <div>
-      <BpkRating
-        className={getClassName('bpk-rating-story')}
-        ariaLabel="9 Excellent would recommend"
-        title="Excellent"
-        subtitle="This place was amazing"
-        value={9}
-        vertical
-      />
-      <br />
-      <BpkRating
-        className={getClassName('bpk-rating-story')}
-        ariaLabel="6.7 Average might recommend"
-        title="Average"
-        subtitle="Might recommend"
-        value={6.7}
-        vertical
-      />
-      <br />
-      <BpkRating
-        className={getClassName('bpk-rating-story')}
-        ariaLabel="2.3 Bad avoid here"
-        title="Bad"
-        subtitle="Avoid here"
-        value={2.3}
-        vertical
-      />
-    </div>
-  ))
-  .add('Large size ratings', () => (
-    <div>
-      <BpkRating
-        ariaLabel="9 Excellent would recommend"
-        title="Excellent"
-        subtitle="This place was amazing"
-        value={9}
-        size={RATING_SIZES.lg}
-      />
-      <br />
-      <BpkRating
-        ariaLabel="6.7 Average might recommend"
-        title="Average"
-        subtitle="Might recommend"
-        value={6.7}
-        size={RATING_SIZES.lg}
-      />
-      <br />
-      <BpkRating
-        ariaLabel="2.3 Bad avoid here"
-        title="Bad"
-        subtitle="Avoid here"
-        value={2.3}
-        size={RATING_SIZES.lg}
-      />
-    </div>
-  ))
-  .add('Large vertical ratings', () => (
-    <div>
-      <BpkRating
-        className={getClassName('bpk-rating-story')}
-        ariaLabel="9 Excellent would recommend"
-        title="Excellent"
-        subtitle="This place was amazing"
-        value={9}
-        size={RATING_SIZES.lg}
-        vertical
-      />
-      <br />
-      <BpkRating
-        className={getClassName('bpk-rating-story')}
-        ariaLabel="6.7 Average might recommend"
-        title="Average"
-        subtitle="Might recommend"
-        value={6.7}
-        size={RATING_SIZES.lg}
-        vertical
-      />
-      <br />
-      <BpkRating
-        className={getClassName('bpk-rating-story')}
-        ariaLabel="2.3 Bad avoid here"
-        title="Bad"
-        subtitle="Avoid here"
-        value={2.3}
-        size={RATING_SIZES.lg}
-        vertical
-      />
-    </div>
-  ))
-  .add('Small size ratings', () => (
-    <div>
-      <BpkRating
-        ariaLabel="9 Excellent would recommend"
-        title="Excellent"
-        subtitle="This place was amazing"
-        value={9}
-        size={RATING_SIZES.sm}
-      />
-      <br />
-      <BpkRating
-        ariaLabel="6.7 Average might recommend"
-        title="Average"
-        subtitle="Might recommend"
-        value={6.7}
-        size={RATING_SIZES.sm}
-      />
-      <br />
-      <BpkRating
-        ariaLabel="2.3 Bad avoid here"
-        title="Bad"
-        subtitle="Avoid here"
-        value={2.3}
-        size={RATING_SIZES.sm}
-      />
-    </div>
-  ))
-  .add('Small vertical ratings', () => (
-    <div>
-      <BpkRating
-        className={getClassName('bpk-rating-story')}
-        ariaLabel="9 Excellent would recommend"
-        title="Excellent"
-        subtitle="This place was amazing"
-        value={9}
-        size={RATING_SIZES.sm}
-        vertical
-      />
-      <br />
-      <BpkRating
-        className={getClassName('bpk-rating-story')}
-        ariaLabel="6.7 Average might recommend"
-        title="Average"
-        subtitle="Might recommend"
-        value={6.7}
-        size={RATING_SIZES.sm}
-        vertical
-      />
-      <br />
-      <BpkRating
-        className={getClassName('bpk-rating-story')}
-        ariaLabel="2.3 Bad avoid here"
-        title="Bad"
-        subtitle="Avoid here"
-        value={2.3}
-        size={RATING_SIZES.sm}
-        vertical
-      />
-    </div>
-  ))
-  .add('Title only ratings', () => (
-    <div>
-      <BpkRating
-        ariaLabel="9 Excellent would recommend"
-        title="Excellent"
-        value={9}
-        size={RATING_SIZES.sm}
-      />
-      <br />
-      <BpkRating
-        ariaLabel="6.7 Average might recommend"
-        title="Average"
-        value={6.7}
-      />
-      <br />
-      <BpkRating
-        ariaLabel="2.3 Bad avoid here"
-        title="Bad"
-        value={2.3}
-        size={RATING_SIZES.lg}
-      />
-      <br />
-      <BpkRating
-        className={getClassName('bpk-rating-story')}
-        ariaLabel="9 Excellent would recommend"
-        title="Excellent"
-        value={9}
-        size={RATING_SIZES.sm}
-        vertical
-      />
-      <br />
-      <BpkRating
-        className={getClassName('bpk-rating-story')}
-        ariaLabel="6.7 Average might recommend"
-        title="Average"
-        value={6.7}
-        vertical
-      />
-      <br />
-      <BpkRating
-        className={getClassName('bpk-rating-story')}
-        ariaLabel="2.3 Bad avoid here"
-        title="Bad"
-        value={2.3}
-        size={RATING_SIZES.lg}
-        vertical
-      />
-    </div>
-  ))
-  .add('Pill ratings', () => (
-    <div>
-      <BpkRating
-        ariaLabel="9 Excellent would recommend"
-        title="Excellent"
-        subtitle="(4,000 reviews)"
-        value={9}
-        size={RATING_SIZES.sm}
-        type={RATING_TYPES.pill}
-      />
-      <br />
-      <BpkRating
-        ariaLabel="6.7 Average might recommend"
-        title="Average"
-        subtitle="(50 reviews)"
-        value={6.7}
-        type={RATING_TYPES.pill}
-      />
-      <br />
-      <BpkRating
-        ariaLabel="2.3 Bad avoid here"
-        title="Bad"
-        subtitle="(1,000 reviews)"
-        value={2.3}
-        size={RATING_SIZES.lg}
-        type={RATING_TYPES.pill}
-      />
-      <br />
-    </div>
-  ))
-  .add('Pill title only ratings', () => (
-    <div>
-      <BpkRating
-        ariaLabel="9 Excellent would recommend"
-        title="Excellent"
-        value={9}
-        size={RATING_SIZES.sm}
-        type={RATING_TYPES.pill}
-      />
-      <br />
-      <BpkRating
-        ariaLabel="6.7 Average might recommend"
-        title="Average"
-        value={6.7}
-        type={RATING_TYPES.pill}
-      />
-      <br />
-      <BpkRating
-        ariaLabel="2.3 Bad avoid here"
-        title="Bad"
-        value={2.3}
-        size={RATING_SIZES.lg}
-        type={RATING_TYPES.pill}
-      />
-      <br />
-    </div>
-  ));
+  .add('Default', DefaultExample)
+  .add('Vertical', Vertical)
+  .add('Large size ratings', LargeSize)
+  .add('Large vertical ratings', LargeVertical)
+  .add('Small size ratings', SmallSize)
+  .add('Small vertical ratings', SmallVertical)
+  .add('Title only ratings', TitleOnly)
+  .add('Pill ratings', PillRatings)
+  .add('Pill title only ratings', PillTitleOnly);
