+++ packages/bpk-component-textarea/stories.js
@@ -18,52 +18,17 @@
 
 /* @flow strict */
 
-import React from 'react';
 import { storiesOf } from '@storybook/react';
-import { action } from 'bpk-storybook-utils';
 
-import BpkTextarea from './index';
-
-const loremIpsum = `Lorem ipsum dolor sit amet, consectetur adipisicing elit. Voluptate repellat assumenda
-necessitatibus reiciendis, porro temporibus expedita excepturi! Nostrum pariatur odit porro, dolorem dignissimos
-laudantium quis, tempore iste non, nam magnam.`;
+import {
+  DefaultExample,
+  PlaceholderExample,
+  DisabledExample,
+  InvalidExample,
+} from './examples';
 
 storiesOf('bpk-component-textarea', module)
-  .add('Default', () => (
-    <BpkTextarea
-      id="default"
-      name="default"
-      value={loremIpsum}
-      onChange={action('input changed')}
-      placeholder="Please enter some text"
-    />
-  ))
-  .add('Placeholder', () => (
-    <BpkTextarea
-      id="placeholder"
-      name="placeholder"
-      value=""
-      onChange={action('input changed')}
-      placeholder="Please enter some text"
-    />
-  ))
-  .add('Disabled', () => (
-    <BpkTextarea
-      id="disabled"
-      name="disabled"
-      value=""
-      onChange={action('input changed')}
-      placeholder="Please enter some text"
-      disabled
-    />
-  ))
-  .add('Invalid', () => (
-    <BpkTextarea
-      id="disabled"
-      name="disabled"
-      value=""
-      onChange={action('input changed')}
-      placeholder="Please enter some text"
-      valid={false}
-    />
-  ));
+  .add('Default', DefaultExample)
+  .add('Placeholder', PlaceholderExample)
+  .add('Disabled', DisabledExample)
+  .add('Invalid', InvalidExample);
