--- packages/bpk-component-tile/README.md
@@ -1,12 +1,0 @@
-# bpk-component-tile
-
-> Backpack tile component.
-
-## Installation
-
-```sh
-npm install bpk-component-tile --save-dev
-```
-
-## Usage
-TODO
