--- packages/bpk-component-tile/package.json
@@ -1,27 +1,0 @@
-{
-  "name": "bpk-component-tile",
-  "version": "0.0.248",
-  "description": "Backpack card component.",
-  "license": "Apache-2.0",
-  "private": true,
-  "repository": {
-    "type": "git",
-    "url": "git@github.com:Skyscanner/backpack.git"
-  },
-  "author": "Backpack Design System <backpack@skyscanner.net>",
-  "main": "index.js",
-  "publishConfig": {
-    "registry": "https://registry.npmjs.org/"
-  },
-  "dependencies": {
-    "bpk-component-icon": "^8.3.17",
-    "bpk-mixins": "^20.1.25",
-    "prop-types": "^15.6.2"
-  },
-  "peerDependencies": {
-    "react": "^16.0.0"
-  },
-  "devDependencies": {
-    "bpk-react-utils": "^3.0.11"
-  }
-}
