--- packages/bpk-component-tile/src/BpkTile.css
@@ -1,1 +1,0 @@
-@keyframes bpk-keyframe-spin{100%{transform:rotate(1turn)}}.bpk-tile{display:flex;height:18.75rem;flex-direction:column;justify-content:flex-end;border-radius:.375rem;background-repeat:no-repeat;background-position:center center;background-size:cover;overflow:hidden;box-shadow:0px 1px 3px 0px rgba(37,32,31,0.3)}.bpk-no-touch-support .bpk-tile:hover:not(:active):not(:disabled){box-shadow:0px 4px 14px 0px rgba(37,32,31,0.25)}:global(.bpk-no-touch-support) .bpk-tile:hover:not(:active):not(:disabled){box-shadow:0px 4px 14px 0px rgba(37,32,31,0.25)}.bpk-tile__illustration{margin:1.5rem;flex:1;background-repeat:no-repeat;background-position:center center;background-size:contain}.bpk-tile__heading{margin:0 0 .375rem 0;padding:0 0 .375rem 0;font-size:1rem;line-height:1.25rem;font-weight:400;box-shadow:0 -1px 0 0 #111236 inset}.bpk-tile__heading--dark{box-shadow:0 -1px 0 0 #fff inset}.bpk-tile__cta{margin:0;font-size:1rem;line-height:1.25rem;font-weight:400}.bpk-tile__cta-icon{fill:#111236}.bpk-tile__cta-icon--dark{fill:#fff}.bpk-tile__scrim{display:flex;min-height:6rem;padding:.75rem;flex-direction:column;justify-content:flex-end;background-image:linear-gradient(-180deg, rgba(255,255,255,0) 0%, #fff 70%);color:#111236}.bpk-tile__scrim--dark{background-image:linear-gradient(-180deg, rgba(17,18,54,0) 0%, rgba(17,18,54,0.4) 70%);color:#fff}
