--- packages/bpk-component-tile/src/BpkTile.scss
@@ -1,93 +1,0 @@
-/*
- * Backpack - Skyscanner's Design System
- *
- * Copyright 2016-2021 Skyscanner Ltd
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-@import '~bpk-mixins/index';
-
-.bpk-tile {
-  display: flex;
-  height: $bpk-one-pixel-rem * 300;
-  flex-direction: column;
-  justify-content: flex-end;
-  border-radius: $bpk-border-radius-sm;
-  background-repeat: no-repeat;
-  background-position: center center;
-  background-size: cover;
-  overflow: hidden;
-
-  @include bpk-box-shadow-sm;
-
-  @include bpk-hover {
-    @include bpk-box-shadow-lg;
-  }
-
-  &__illustration {
-    margin: $bpk-spacing-base;
-    flex: 1;
-    background-repeat: no-repeat;
-    background-position: center center;
-    background-size: contain;
-  }
-
-  &__heading {
-    margin: 0 0 $bpk-spacing-xs 0;
-    padding: 0 0 $bpk-spacing-xs 0;
-
-    @include bpk-text-base;
-    @include bpk-border-bottom-sm($bpk-color-sky-gray);
-
-    &--dark {
-      @include bpk-border-bottom-sm($bpk-color-white);
-    }
-  }
-
-  &__cta {
-    @include bpk-text;
-    @include bpk-text--base;
-  }
-
-  &__cta-icon {
-    fill: $bpk-color-sky-gray;
-
-    &--dark {
-      fill: $bpk-color-white;
-    }
-  }
-
-  &__scrim {
-    display: flex;
-    min-height: $bpk-one-pixel-rem * 96;
-    padding: $bpk-spacing-sm;
-    flex-direction: column;
-    justify-content: flex-end;
-    background-image: linear-gradient(
-      -180deg,
-      rgba($bpk-color-white, 0) 0%,
-      $bpk-color-white 70%
-    );
-    color: $bpk-color-sky-gray;
-
-    &--dark {
-      background-image: linear-gradient(
-        -180deg,
-        rgba($bpk-color-sky-gray, 0) 0%,
-        rgba($bpk-color-sky-gray, 0.4) 70%
-      );
-      color: $bpk-color-white;
-    }
-  }
-}
