+++ packages/bpk-storybook-utils/src/BpkDarkExampleWrapper.css
@@ -1,1 +1,1 @@
-@keyframes bpk-keyframe-spin{100%{transform:rotate(1turn)}}.bpk-dark-example-wrapper{border-radius:.375rem;background-color:#042759;background-color:var(--bpk-docs-sidebar-background, #042759)}
+@keyframes bpk-keyframe-spin{100%{transform:rotate(1turn)}}.bpk-dark-example-wrapper{border-radius:.375rem;background-color:#042759;background-color:var(--bpk-docs-sidebar-background, #042759)}.bpk-dark-example-wrapper--padded{padding:1.125rem}
